chore: move `HyperdriveIndex` into its own file
This change should have no impact.
