+++ src/blob-store/index.js
@@ -1,6 +1,3 @@
-import Hyperdrive from 'hyperdrive'
-import b4a from 'b4a'
-import util from 'node:util'
 import { pipeline } from 'node:stream'
 import { discoveryKey } from 'hypercore-crypto'
 import { Downloader } from './downloader.js'
@@ -8,7 +5,9 @@ import { createEntriesStream } from './entries-stream.js'
 import { FilterEntriesStream } from './utils.js'
 import { noop } from '../utils.js'
 import { TypedEmitter } from 'tiny-typed-emitter'
+import { HyperdriveIndexImpl as HyperdriveIndex } from './hyperdrive-index.js'
 
+/** @import Hyperdrive from 'hyperdrive' */
 /** @import { JsonObject } from 'type-fest' */
 /** @import { Readable as NodeReadable } from 'node:stream' */
 /** @import { Readable as StreamxReadable, Writable } from 'streamx' */
@@ -249,66 +248,6 @@ export class BlobStore extends TypedEmitter {
   }
 }
 
-// Don't want to export the class, but do want to export the type.
-/** @typedef {HyperdriveIndex} THyperdriveIndex */
-
-/**
- * @extends {TypedEmitter<{ 'add-drive': (drive: Hyperdrive) => void }>}
- */
-class HyperdriveIndex extends TypedEmitter {
-  /** @type {Map<string, Hyperdrive>} */
-  #hyperdrives = new Map()
-  #writer
-  #writerKey
-  /** @param {import('../core-manager/index.js').CoreManager} coreManager */
-  constructor(coreManager) {
-    super()
-    /** @type {undefined | Hyperdrive} */
-    let writer
-    const corestore = new PretendCorestore({ coreManager })
-    const blobIndexCores = coreManager.getCores('blobIndex')
-    const writerCoreRecord = coreManager.getWriterCore('blobIndex')
-    this.#writerKey = writerCoreRecord.key
-    for (const { key } of blobIndexCores) {
-      // @ts-ignore - we know pretendCorestore is not actually a Corestore
-      const drive = new Hyperdrive(corestore, key)
-      // We use the discovery key to derive the id for a drive
-      this.#hyperdrives.set(getDiscoveryId(key), drive)
-      if (key.equals(this.#writerKey)) {
-        writer = drive
-      }
-    }
-    if (!writer) {
-      throw new Error('Could not find a writer for the blobIndex namespace')
-    }
-    this.#writer = writer
-
-    coreManager.on('add-core', ({ key, namespace }) => {
-      if (namespace !== 'blobIndex') return
-      // We use the discovery key to derive the id for a drive
-      const driveId = getDiscoveryId(key)
-      if (this.#hyperdrives.has(driveId)) return
-      // @ts-ignore - we know pretendCorestore is not actually a Corestore
-      const drive = new Hyperdrive(corestore, key)
-      this.#hyperdrives.set(driveId, drive)
-      this.emit('add-drive', drive)
-    })
-  }
-  get writer() {
-    return this.#writer
-  }
-  get writerKey() {
-    return this.#writerKey
-  }
-  [Symbol.iterator]() {
-    return this.#hyperdrives.values()
-  }
-  /** @param {string} driveId */
-  get(driveId) {
-    return this.#hyperdrives.get(driveId)
-  }
-}
-
 /**
  * @template {object} T
  * @template {object} U
@@ -332,56 +271,6 @@ function proxyProps(target, props) {
 /** @param {Pick<BlobId, 'type' | 'variant' | 'name'>} opts */
 function makePath({ type, variant, name }) {
   return `/${type}/${variant}/${name}`
-}
-
-/**
- * Implements the `get()` method as used by hyperdrive-next. It returns the
- * relevant cores from the Mapeo CoreManager.
- */
-class PretendCorestore {
-  #coreManager
-  /**
-   * @param {object} options
-   * @param {import('../core-manager/index.js').CoreManager} options.coreManager
-   */
-  constructor({ coreManager }) {
-    this.#coreManager = coreManager
-  }
-
-  /**
-   * @param {Buffer | { publicKey: Buffer } | { name: string }} opts
-   * @returns {import('hypercore')<"binary", Buffer> | undefined}
-   */
-  get(opts) {
-    if (b4a.isBuffer(opts)) {
-      opts = { publicKey: opts }
-    }
-    if ('key' in opts) {
-      // @ts-ignore
-      opts.publicKey = opts.key
-    }
-    if ('publicKey' in opts) {
-      // NB! We should always add blobIndex (Hyperbee) cores to the core manager
-      // before we use them here. We would only reach the addCore path if the
-      // blob core is read from the hyperbee header (before it is added to the
-      // core manager)
-      return (
-        this.#coreManager.getCoreByKey(opts.publicKey) ||
-        this.#coreManager.addCore(opts.publicKey, 'blob').core
-      )
-    } else if (opts.name === 'db') {
-      return this.#coreManager.getWriterCore('blobIndex').core
-    } else if (opts.name.includes('blobs')) {
-      return this.#coreManager.getWriterCore('blob').core
-    } else {
-      throw new Error(
-        'Unsupported corestore.get() with opts ' + util.inspect(opts)
-      )
-    }
-  }
-
-  /** no-op */
-  close() {}
 }
 
 /**
